+++ model/model.h
@@ -31,7 +31,7 @@ public:
 	void setParent(Model* parent) { m_parent = parent; }
 
 	//get the origin of this Object Coordinate Space
-	Vec3f getOrigin();
+	Vec3d getOrigin();
 
 	/**
 	 * \brief Call controller, be called by modelView
